Create ScreensNavigation that provides view model aware builder
