+++ examples/mvi-desktop/src/jvmMain/kotlin/io/github/lukwol/examples/AppNavigation.kt
@@ -7,8 +7,7 @@ import io.github.lukwol.examples.screens.second.SecondScreen
 import io.github.lukwol.examples.screens.second.SecondScreenViewModel
 import io.github.lukwol.examples.screens.third.ThirdScreen
 import io.github.lukwol.examples.screens.third.ThirdScreenViewModel
-import io.github.lukwol.screens.navigation.ScreensNavigation
-import io.github.lukwol.viewmodel.screens.navigation.screen
+import io.github.lukwol.viewmodel.screens.navigation.ScreensNavigation
 
 @Composable
 fun AppNavigation() {
