+++ examples/mvvm-web/src/jsMain/kotlin/io.github.lukwol.examples/AppNavigation.kt
@@ -5,8 +5,7 @@ import io.github.lukwol.examples.screens.first.FirstScreen
 import io.github.lukwol.examples.screens.first.FirstScreenViewModel
 import io.github.lukwol.examples.screens.second.SecondScreen
 import io.github.lukwol.examples.screens.second.SecondScreenViewModel
-import io.github.lukwol.screens.navigation.ScreensNavigation
-import io.github.lukwol.viewmodel.screens.navigation.screen
+import io.github.lukwol.viewmodel.screens.navigation.ScreensNavigation
 
 @Composable
 fun AppNavigation() {
