+++ viewmodel-screens-navigation/src/jvmTest/kotlin/io/github/lukwol/viewmodel/screens/navigation/components/ScreenNavigation.kt
@@ -1,9 +1,8 @@
 package io.github.lukwol.viewmodel.screens.navigation.components
 
 import androidx.compose.runtime.Composable
-import io.github.lukwol.screens.navigation.ScreensNavigation
+import io.github.lukwol.viewmodel.screens.navigation.ScreensNavigation
 import io.github.lukwol.viewmodel.screens.navigation.data.TestRoutes
-import io.github.lukwol.viewmodel.screens.navigation.screen
 
 @Suppress("TestFunctionName")
 @Composable
